Use Case Diagram

+++ src/main/java/com/project/Health_Bot/model/Alimento.java
@@ -60,7 +60,7 @@ public class Alimento {
      * 
      * @return energia
      */
-    public float getKcal() {
+    public int getKcal() {
         return kcal;
     }
